+++ wandb/plot/roc_curve.py
@@ -1,8 +1,9 @@
 from typing import Optional
 
+from utils import test_missing, test_types
+
 import wandb
 from wandb import util
-from wandb.plots.utils import test_missing, test_types
 
 
 def roc_curve(
